Move parseAttribute to attribute.ts
