+++ packages/alfa-dom/src/node/attribute.ts
@@ -1,5 +1,8 @@
+import { Mapper } from "@siteimprove/alfa-mapper";
 import { None, Option } from "@siteimprove/alfa-option";
+import { Parser } from "@siteimprove/alfa-parser";
 import { Predicate } from "@siteimprove/alfa-predicate";
+import { Ok, Result } from "@siteimprove/alfa-result";
 
 import { Namespace } from "../namespace";
 import { Node } from "../node";
@@ -136,6 +139,13 @@ export namespace Attribute {
       attribute.value,
       owner
     );
+  }
+
+  export function parseAttribute<RESULT, ERROR>(
+    parser: Parser<string, RESULT, ERROR>
+  ): Mapper<Attribute, Result<RESULT, ERROR>> {
+    return (attribute) =>
+      parser(attribute.value).andThen(([_, value]) => Ok.of(value));
   }
 }
