+++ packages/alfa-dom/test/node/parse-attribute.spec.ts
@@ -4,11 +4,10 @@ import { EnumeratedValueError, Parser } from "@siteimprove/alfa-parser";
 import { Err, Ok } from "@siteimprove/alfa-result";
 import { test } from "@siteimprove/alfa-test";
 
-import { Attribute, Element } from "../../src";
-import {
-  parseAttribute,
-  parseEnumeratedAttribute,
-} from "../../src/tables/helpers";
+import { Attribute, Element } from "../../src/index";
+import { parseEnumeratedAttribute } from "../../src/tables/helpers";
+
+const { parseAttribute } = Attribute;
 
 const makeAttribute = (str: string) => Attribute.of(None, None, "dummy", str);
